+++ src/components/Faucet/Hero.tsx
@@ -35,7 +35,7 @@ export default function Hero({
     metaMaskWalletIdConnectHandler,
     walletLinked,
     projects,
-    walletLinkUrl,
+    walletAuthUrl,
   } = useContext(MetamaskProviderContext)
   const isExtensionActive = sdk.isExtensionActive()
   const [isWalletLinking, setIsWalletLinking] = useState(false)
@@ -65,7 +65,7 @@ export default function Hero({
       responseMsg: null,
       timestamp: Date.now(),
     })
-    window.location.href = walletLinkUrl
+    window.location.href = walletAuthUrl
   }
 
   const handleRequestEth = () => {
@@ -120,11 +120,9 @@ export default function Hero({
                 label="Wallet address"
                 disabled={isLoading}
                 value={inputValue}
-                className={styles.input}
-                placeholder="ex. 0x"
+                placeholder="ex. 0x or ENS"
                 onChange={handleOnInputChange}
               />
-              {/* isExtensionActive && Object.keys(projects).length */}
               <div className={clsx(!!Object.keys(projects).length && styles.alignedButtons)}>
                 <Button
                   as="button"