+++ src/lib/data.js
@@ -116,9 +116,19 @@ export const NETWORK_LINKS = [
     logo: `${LOGO_PATH}/logo-polygon.svg`,
   },
   {
+    name: "Scroll",
+    href: "/services/reference/scroll",
+    logo: `${LOGO_PATH}/logo-scroll.svg`,
+  },
+  {
     name: "Starknet",
     href: "/services/reference/starknet",
     logo: `${LOGO_PATH}/logo-starknet.svg`,
+  },
+  {
+    name: "Unichain",
+    href: "/services/reference/unichain",
+    logo: `${LOGO_PATH}/logo-unichain.svg`,
   },
   {
     name: "ZKsync Era",