+++ docs/whats-new.md
@@ -9,11 +9,30 @@ The latest major MetaMask documentation updates are listed by the month they wer
 For a comprehensive list of recent product changes, visit the "Release Notes" section at the bottom
 of the [MetaMask developer page](https://metamask.io/developer/).
 
+## November 2024
+
+- Documented [Unichain Sepolia](/services/reference/unichain) support. ([#1725](https://github.com/MetaMask/metamask-docs/pull/1725))
+- Updated Snaps [Custom UI documentation](/snaps/features/custom-ui/) for MetaMask Extension version 12.6.
+  ([#1715](https://github.com/MetaMask/metamask-docs/pull/1715))
+- Added tutorial for
+  [creating a simple Starknet dapp](/wallet/how-to/use-non-evm-networks/starknet/create-a-simple-starknet-dapp).
+  ([#1656](https://github.com/MetaMask/metamask-docs/pull/1656))
+- Documented Snaps [`Card`](/snaps/features/custom-ui/#card) and [`Section`](/snaps/features/custom-ui/#section) UI components.
+  ([#1707](https://github.com/MetaMask/metamask-docs/pull/1707)) ([#1700](https://github.com/MetaMask/metamask-docs/pull/1700))
+
 ## October 2024
 
+- Documented support for [non-EVM networks](/wallet/how-to/use-non-evm-networks), starting with [Starknet](/wallet/how-to/use-non-evm-networks/starknet).
+  ([#1509](https://github.com/MetaMask/metamask-docs/pull/1509))
+- Documented [Scroll](/services/reference/scroll) support.
+  ([#1641](https://github.com/MetaMask/metamask-docs/pull/1641))
 - Updated [Wallet landing page](/wallet) and added [Connect to MetaMask](/wallet/connect) section
   with SDK, third-party libraries, and Wallet API connection options.
   ([#1494](https://github.com/MetaMask/metamask-docs/pull/1494))
+- Documented [`snap_getPreferences`](/snaps/reference/snaps-api/#snap_getpreferences).
+  ([#1681](https://github.com/MetaMask/metamask-docs/pull/1681))
+- Documented [Snaps custom UI dialogs](/snaps/features/custom-ui/dialogs).
+  ([#1682](https://github.com/MetaMask/metamask-docs/pull/1682))
 
 ## September 2024